pyproject.toml
[project]
name = "sync-ai"
version = "0.1.0"
description = "AI Task Manager"
authors = [
    {name = "Benjamin Delacruz",email = "bydelacruz@icloud.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "groq (>=1.0.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "bcrypt (==4.0.1)",
    "alembic (>=1.18.3,<2.0.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
